hw10

+++ src/test/java/edu/hw10/Task2Test/Task2Test.java
@@ -81,7 +81,7 @@ public class Task2Test {
     @AfterAll
     public static void deleteCacheFiles() {
         File[] files = cacheDir.listFiles();
-        if (files != null || files.length >= 1) {
+        if (files != null && files.length >= 1) {
             for (File file : files) {
                 if (!file.isDirectory()) {
                     file.delete();
